Refactor Metal renderer initialization and resource management
- Updated GPU.cpp to map RENDERID constants to core3DList indices for improved renderer selection.
- Enhanced MetalRender.mm to verify the existence of the default library in SharedMetalData before accessing it, with fallback handling.
- Improved MetalShaders.metal by explicitly casting fogIndex to uint for clarity.
- Modified meson.build to ensure Metal shaders are built by default.
- Expanded metal_bootstrap.mm to include robust library loading from various paths and added a new initialization function for embedded data.

+++ desmume/src/MetalShaders.metal
@@ -297,7 +297,8 @@ fragment float4 fogFragment(
     
     // Look up fog density from the 32-entry table
     // Density is normalized [0, 1] where 1 = full fog
-    float fogDensity = fogDensityTable.read(fogIndex).r;
+    // Cast to uint explicitly to resolve ambiguity
+    float fogDensity = fogDensityTable.read(uint(fogIndex)).r;
     
     // Return fog weight as RGBA
     // This will be blended with the scene using special blend modes
